Separate lib_watch.go
